Peer ID is short

+++ src/ReadyM.Api.Multiplayer/Protocol/CustomEventHeader.cs
@@ -3,18 +3,18 @@
 namespace ReadyM.Api.Multiplayer.Protocol;
 
 /// <summary>
-/// Spans 4 bytes in standard relay modes and (6 + 4 * peers) in peer-targeted relay mode.
+/// Spans 4 bytes in standard relay modes and (6 + 2 * peers) in peer-targeted relay mode.
 /// - 1 byte for eventCode
 /// - 2 bytes for sender
 /// - 1 byte for relayMode and eventCaching (bit packed)
 /// - 2 bytes for peers count
-/// - peers * 4 bytes for peers
+/// - peers * 2 bytes for peers
 /// </summary>
-public readonly struct CustomEventHeader(byte eventCode, short sender, int[]? peers, RelayMode relayMode = RelayMode.Others, EventCaching eventCaching = EventCaching.DoNotCache)
+public readonly struct CustomEventHeader(byte eventCode, short sender, short[]? peers, RelayMode relayMode = RelayMode.Others, EventCaching eventCaching = EventCaching.DoNotCache)
 {
     public readonly byte EventCode = eventCode;
     public readonly short Sender = sender;
     public readonly RelayMode RelayMode = relayMode;
     public readonly EventCaching EventCaching = eventCaching;
-    public readonly int[]? Peers = peers;
+    public readonly short[]? Peers = peers;
 }
